fix(ssi): create injector interface and refactor webhooks
This commit adds an injector interface and refactors the tagsfromlabels
and the config webhooks to use it. This will allow us to use the
injector logic without needing to instantiate a webhook.
